Move NativeModule into modules dir
