+++ src/js/module/NativeModule.js
